refactor so lobby status is calculated on frontend
