+++ frontend/src/organisms/LobbyPane.tsx
@@ -1,0 +1,34 @@
+import { Box, Button, List, Typography } from '@mui/material';
+import PlayerListItem from '../molecules/PlayerListItem.tsx';
+import { GamePageProps } from '../pages/GamePage.tsx';
+
+type LobbyPaneProps = GamePageProps & {
+  playerRenameEvent: (newName: string) => void;
+  startGameEvent: () => void;
+};
+function LobbyPane({ gameState, playerRenameEvent, startGameEvent }: LobbyPaneProps) {
+  const isCreator = gameState.creator === localStorage.getItem('playerId');
+
+  return (
+    <Box>
+      <List>
+        {gameState.players.map((player) => (
+          <PlayerListItem
+            key={player.id}
+            player={player}
+            rename={playerRenameEvent}
+            creator={gameState.creator}
+          />
+        ))}
+      </List>
+      {isCreator && (
+        <Button onClick={startGameEvent} disabled={gameState.players.length < 3}>
+          Start Game
+        </Button>
+      )}
+      {!isCreator && <Typography>Waiting for Host...</Typography>}
+    </Box>
+  );
+}
+
+export default LobbyPane;
